# openapi.yaml
openapi: 3.1.0
info:
  title: Tateca API
  description: "Group expense management API with multi-currency support. Track expenses,\
    \ loans, and settlements across multiple users and currencies."
  contact:
    name: Tateca Team
    url: https://tateca.net
    email: support@tateca.net
  license:
    name: Proprietary
    url: https://tateca.net/terms
  version: 1.0.0
...
servers:
- url: http://localhost:8080
  description: Local Development
- url: https://staging-api.tateca.net
  description: Staging
- url: https://api.tateca.net
  description: Production
security:
- bearerAuth: []
...
paths:
...
  /exchange-rate/{date}:
    get:
      tags:
      - Exchange Rates
      summary: Get exchange rates for a specific date
      description: Retrieve exchange rates for a specific date in ISO 8601 format
        (YYYY-MM-DD)
      operationId: getExchangeRate
      parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: Exchange rates retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateResponseDTO"
        "404":
          description: Exchange rates not found for the specified date
          content:
            application/json:
              example:
                status: 404
                message: Exchange rates not found for date
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                status: 500
                message: Database connection error
        "400":
          description: Validation error - Invalid date format
          content:
            application/json:
              examples:
                Invalid date format:
                  description: "When date is not in ISO 8601 format (YYYY-MM-DD) or\
                    \ invalid date (e.g., 2024-13-45)"
                  value:
                    status: 400
                    message: "Invalid date format: expected ISO 8601 date (YYYY-MM-DD)"
                Malformed date string:
                  description: When date parameter is not a valid date string
                  value:
                    status: 400
                    message: Failed to convert value of type 'java.lang.String' to
                      required type 'java.time.LocalDate'
        "401":
          description: Unauthorized - Invalid or missing Firebase JWT token
          content:
            application/json:
              example:
                status: 401
                message: Unauthorized
...
components:
  schemas:
...
    ExchangeRateResponse:
      type: object
      description: Exchange rate information for a currency
      properties:
        currency_code:
          type: string
          description: ISO 4217 currency code
          example: USD
        jp_currency_name:
          type: string
          description: Currency name in Japanese
          example: 米ドル
        eng_currency_name:
          type: string
          description: Currency name in English
          example: US Dollar
        jp_country_name:
          type: string
          description: Country name in Japanese
          example: アメリカ合衆国
        eng_country_name:
          type: string
          description: Country name in English
          example: United States
        currency_symbol:
          type: string
          description: Currency symbol
          example: $
        symbol_position:
          type: string
          description: Position of currency symbol
          enum:
          - PREFIX
          - SUFFIX
        exchange_rate:
          type: string
          description: Exchange rate to JPY
          example: 150.25
...
    ExchangeRateResponseDTO:
      type: object
      description: Exchange rate list response
      properties:
        exchange_rate:
          type: array
          description: List of exchange rates
          items:
            $ref: "#/components/schemas/ExchangeRateResponse"
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Firebase JWT token for user authentication
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      description: API Key for internal system operations (EventBridge + Lambda)
      name: X-API-Key
      in: header