# src/main/java/com/tateca/tatecabackend/controller/ExchangeRateController.java
package com.tateca.tatecabackend.controller;

import com.tateca.tatecabackend.dto.response.ExchangeRateResponseDTO;
import com.tateca.tatecabackend.service.ExchangeRateService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.time.LocalDate;

@RestController
@RequestMapping(value = "/exchange-rate", produces = MediaType.APPLICATION_JSON_VALUE)
@RequiredArgsConstructor
@Tag(name = "Exchange Rates", description = "Currency exchange rate operations")
public class ExchangeRateController {
    private final ExchangeRateService service;

    @GetMapping("/{date}")
    @Operation(summary = "Get exchange rates for a specific date", description = "Retrieve exchange rates for a specific date in ISO 8601 format (YYYY-MM-DD)")
    @ApiResponses({
        @ApiResponse(
            responseCode = "200",
            description = "Exchange rates retrieved successfully"
        ),
        @ApiResponse(
            responseCode = "400",
            description = "Validation error - Invalid date format",
            content = @Content(
                mediaType = "application/json",
                examples = {
                    @ExampleObject(
                        name = "Invalid date format",
                        description = "When date is not in ISO 8601 format (YYYY-MM-DD) or invalid date (e.g., 2024-13-45)",
                        value = """
                            {
                              "status": 400,
                              "message": "Invalid date format: expected ISO 8601 date (YYYY-MM-DD)"
                            }
                            """
                    ),
                    @ExampleObject(
                        name = "Malformed date string",
                        description = "When date parameter is not a valid date string",
                        value = """
                            {
                              "status": 400,
                              "message": "Failed to convert value of type 'java.lang.String' to required type 'java.time.LocalDate'"
                            }
                            """
                    )
                }
            )
        ),
        @ApiResponse(
            responseCode = "401",
            description = "Unauthorized - Invalid or missing Firebase JWT token",
            content = @Content(
                mediaType = "application/json",
                examples = @ExampleObject(
                    value = """
                        {
                          "status": 401,
                          "message": "Unauthorized"
                        }
                        """
                )
            )
        ),
        @ApiResponse(
            responseCode = "404",
            description = "Exchange rates not found for the specified date",
            content = @Content(
                mediaType = "application/json",
                examples = @ExampleObject(
                    value = """
                        {
                          "status": 404,
                          "message": "Exchange rates not found for date"
                        }
                        """
                )
            )
        ),
        @ApiResponse(
            responseCode = "500",
            description = "Internal Server Error",
            content = @Content(
                mediaType = "application/json",
                examples = @ExampleObject(
                    value = """
                        {
                          "status": 500,
                          "message": "Database connection error"
                        }
                        """
                )
            )
        )
    })
    public ResponseEntity<ExchangeRateResponseDTO> getExchangeRate(
            @PathVariable("date") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date
    ) {
        return ResponseEntity.ok(service.getExchangeRate(date));
    }
}